Fix deployment targets
- Lower iOS and tvOS to 9.0, closes #16
- Correct SPM platforms and make consistent with podspec, closes #18

This also separates tests into `ObservationTests`, which we can mark as `@available` so that building/running via SPM works.
