+++ Foil.podspec
@@ -14,8 +14,8 @@ Pod::Spec.new do |s|
 
    s.swift_version = '5.3'
 
-   s.ios.deployment_target = '12.0'
-   s.tvos.deployment_target = '12.0'
+   s.ios.deployment_target = '9.0'
+   s.tvos.deployment_target = '9.0'
    s.watchos.deployment_target = '5.0'
    s.osx.deployment_target = '10.13'
