+++ Tests/IntegrationTests.swift
@@ -11,14 +11,12 @@
 //  Copyright © 2021-present Jesse Squires
 //
 
-import Combine
 @testable import Foil
 import XCTest
 
 final class IntegrationTests: XCTestCase {
 
     let settings = TestSettings()
-    var cancellable = Set<AnyCancellable>()
 
     func test_Integration_Bool() {
         let defaultValue = settings.flag
@@ -159,23 +157,5 @@ final class IntegrationTests: XCTestCase {
 
         let expectedValue = ["key1": TestFruit.apple.rawValue, "key2": TestFruit.orange.rawValue]
         XCTAssertEqual(TestSettings.store.fetch("customRawRepresented"), expectedValue)
-    }
-
-    func test_Integration_Publisher() {
-        let promise = expectation(description: #function)
-        var publishedValue: String?
-
-        settings
-            .publisher(for: \.nickname, options: [.new])
-            .sink {
-                publishedValue = $0
-                promise.fulfill()
-            }
-            .store(in: &cancellable)
-
-        settings.nickname = "abc123"
-        wait(for: [promise], timeout: 5)
-
-        XCTAssertEqual(settings.nickname, publishedValue)
     }
 }
